Update unit tests workflow and install script

+++ phpunit/install.sh
@@ -19,6 +19,30 @@ WP_CORE_DIR=${WP_CORE_DIR-$TMPDIR/wordpress/}
 RETRIES=5
 SLEEP=10
 
+download_and_extract_wordpress_develop() {
+	local ref=$1
+	local archive_url=$2
+	local dest_dir=$3
+
+	rm -rf "$dest_dir"
+	mkdir -p "$dest_dir"
+
+	local archive_file
+	archive_file=$(mktemp)
+	if ! retry_run curl -sSL "$archive_url" -o "$archive_file"; then
+		rm -f "$archive_file"
+		return 1
+	fi
+
+	if ! tar -xzf "$archive_file" -C "$dest_dir" --strip-components=1; then
+		rm -f "$archive_file"
+		return 1
+	fi
+
+	rm -f "$archive_file"
+	return 0
+}
+
 # retry_run: run a command and retry on failure
 retry_run() {
 	local n=0
@@ -58,25 +82,55 @@ fi
 set -ex
 
 install_wp_and_test_suite() {
+	local wp_develop_dir="$TMPDIR/wordpress-develop"
+	local wp_develop_url=""
+	local wp_develop_ref=""
+
+	if [[ $WP_TESTS_TAG == trunk ]]; then
+		wp_develop_ref="trunk"
+		wp_develop_url="https://codeload.github.com/WordPress/wordpress-develop/tar.gz/${wp_develop_ref}"
+	elif [[ $WP_TESTS_TAG =~ ^branches/ ]]; then
+		wp_develop_ref=${WP_TESTS_TAG#branches/}
+		wp_develop_url="https://codeload.github.com/WordPress/wordpress-develop/tar.gz/${wp_develop_ref}"
+	elif [[ $WP_TESTS_TAG =~ ^tags/ ]]; then
+		wp_develop_ref=${WP_TESTS_TAG#tags/}
+		wp_develop_url="https://codeload.github.com/WordPress/wordpress-develop/tar.gz/${wp_develop_ref}"
+	fi
+
 	# setup up WordPress
-	if [ ! -d $WP_CORE_DIR ]; then
-		mkdir -p $WP_CORE_DIR
-		# Retry svn checkout in case of transient network issues
-		retry_run svn co --quiet https://develop.svn.wordpress.org/${WP_TESTS_TAG}/src/ $WP_CORE_DIR
+	if [ ! -d "$WP_CORE_DIR" ]; then
+		mkdir -p "$WP_CORE_DIR"
+		if [[ -n "$wp_develop_url" ]] && download_and_extract_wordpress_develop "$wp_develop_ref" "$wp_develop_url" "$wp_develop_dir"; then
+			rm -rf "$WP_CORE_DIR"
+			mkdir -p "$WP_CORE_DIR"
+			cp -R "$wp_develop_dir/src/." "$WP_CORE_DIR"
+		else
+			# Fallback to SVN if GitHub is unavailable.
+			retry_run svn co --quiet https://develop.svn.wordpress.org/${WP_TESTS_TAG}/src/ "$WP_CORE_DIR"
+		fi
 	fi
 
 	# set up testing suite if it doesn't yet exist
-	if [ ! -d $WP_TESTS_DIR ]; then
+	if [ ! -d "$WP_TESTS_DIR" ]; then
 		# set up testing suite
-		mkdir -p $WP_TESTS_DIR
-		# Retry svn checkouts for includes and data
-		retry_run svn co --quiet https://develop.svn.wordpress.org/${WP_TESTS_TAG}/tests/phpunit/includes/ $WP_TESTS_DIR/includes
-		retry_run svn co --quiet https://develop.svn.wordpress.org/${WP_TESTS_TAG}/tests/phpunit/data/ $WP_TESTS_DIR/data
+		mkdir -p "$WP_TESTS_DIR"
+		if [[ -n "$wp_develop_url" ]] && [ -d "$wp_develop_dir/tests/phpunit" ]; then
+			cp -R "$wp_develop_dir/tests/phpunit/includes" "$WP_TESTS_DIR/includes"
+			cp -R "$wp_develop_dir/tests/phpunit/data" "$WP_TESTS_DIR/data"
+		else
+			# Fallback to SVN if GitHub is unavailable.
+			retry_run svn co --quiet https://develop.svn.wordpress.org/${WP_TESTS_TAG}/tests/phpunit/includes/ "$WP_TESTS_DIR/includes"
+			retry_run svn co --quiet https://develop.svn.wordpress.org/${WP_TESTS_TAG}/tests/phpunit/data/ "$WP_TESTS_DIR/data"
+		fi
 	fi
 
 	if [ ! -f wp-tests-config.php ]; then
 		# download wp-tests-config-sample.php with retries
-		retry_run curl -sSL https://develop.svn.wordpress.org/${WP_TESTS_TAG}/wp-tests-config-sample.php -o "$WP_TESTS_DIR/wp-tests-config.php"
+		if [ -f "$wp_develop_dir/wp-tests-config-sample.php" ]; then
+			cp "$wp_develop_dir/wp-tests-config-sample.php" "$WP_TESTS_DIR/wp-tests-config.php"
+		else
+			retry_run curl -sSL https://develop.svn.wordpress.org/${WP_TESTS_TAG}/wp-tests-config-sample.php -o "$WP_TESTS_DIR/wp-tests-config.php"
+		fi
 		# remove all forward slashes in the end
 		WP_CORE_DIR=$(echo $WP_CORE_DIR | sed "s:/\+$::")
 		sed -i "s:dirname( __FILE__ ) . '/src/':'$WP_CORE_DIR/':" "$WP_TESTS_DIR"/wp-tests-config.php
